requirements.txt
# Web Server & Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database & ORM
sqlmodel==0.0.14
pyodbc==5.0.1
pymysql==1.0.3

# Google Cloud & API
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utilities
python-dotenv==1.0.0
pymupdf==1.23.0
yfinance
camelot-py[cv]==1.0.9